TINY-8417: Code cleanup from PR and adding missing find logic
