+++ modules/alloy/CHANGELOG.md
@@ -6,9 +6,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-### Fixed
-- Menu buttons now get an "aria-controls" attribute rather than an "aria-owns" attribute when activated.
-
 ### Added
 - Added a new `buildOrPatch` function to the `SystemApi` to allow reusing the passed obsoleted dom element.
 
@@ -18,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Upgraded to Katamari 9.0, which includes breaking changes to the `Optional` API used in this module.
+- Dropdowns now use an "aria-controls" attribute rather than an "aria-owns" attribute when activated.
 - Firefox will now use the native `focusin` and `focusout` events instead of capturing the `focus` and `blur` events.
 - Horizontal and Vertical slider data is now just a number, instead of objects with single `x` or `y` properties #TINY-8304
