+++ modules/alloy/src/test/ts/browser/ui/button/FloatingToolbarButtonTest.ts
@@ -91,7 +91,7 @@ UnitTest.asynctest('FloatingToolbarButtonTest', (success, failure) => {
           children: [
             s.element('div', {
               attrs: {
-                id: str.contains('aria-owns')
+                id: str.contains('aria-controls')
               },
               children: [
                 s.element('div', {
