+++ modules/tinymce/CHANGELOG.md
@@ -88,6 +88,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deleting content would sometimes not fire `beforeinput` and `input` events as expected #TINY-8168  #TINY-8329
 - Alignment would sometimes be removed on parent elements when changing alignment on certain inline nodes, such as images #TINY-8308
 - Anchor elements would render incorrectly when using the `allow_html_in_named_anchor` option #TINY-3799
+- Fixed sub-menu items not read by screen readers. Patch contributed by westonkd #TINY-8417
 
 ### Removed
 - Removed the jQuery integration #TINY-4518
